requirements.txt
altair==5.1.2
numpy==1.26.1
pandas==2.1.2
pydeck==0.8.0
streamlit==1.28.1
googletrans
gtts
openai-whisper
SpeechRecognition
ffmpeg